requirements.txt
# flask stuff
flask==3.0.0
flask_cors==4.0.0
flask_babel==3.1.0
Flask-WTF==1.2.1
# Google Stuff
firebase_admin==6.3.0
google-api-python-client==2.110.0
google-cloud-secret-manager==2.17.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-cloud-logging==3.9.0
google-cloud-bigquery==3.18.0
werkzeug==3.0.0
waitress==2.1.2
oauthlib==3.2.2
requests==2.28.1
requests-oauthlib==1.3.1
pyotp==2.9.0
oauth2client==4.1.3
# Okta authentication
# okta-jwt-verifier==1.3.1
# okta-oidc==1.5.0
# library for qrcode
qrcode==7.4.2
# additional details
pytz==2023.3.post1
BeautifulSoup4==4.12.2
pillow==10.1.0
